Fix potential precision issue with phase in Kane-Mele drivers
Following the recent testing work on the KM nanoflakes, I noticed we can
potentially encounter numerical precision problems when reading the PHI
value from input files: the default value for phi is expressed in source
as PI/2d0, where PI is defined in SF_CONSTANTS to a good precision. But
whenever a used.input file is written such value is truncated to just 10
digits, potentially hindering final accuracy, for every computation that
actually reads the value from a file (i.e. basically every single one).

The solution is to pass a /differently-named/ input parameter, defining
the phase in units of PI and let the driver multiply this number to the
PI value defined in SciFortran. For typical fractions of PI that would
appear in our work this would maintain the full precision.

The new input parameter has a new name so to avoid reading wrong values
from old input files (e.g. reading PHI=1.570796327 in the file and then
multiplying again by PI), so that if you parse an old input file you'll
just get the default NPHI=0.5d0 value and find that in the used.input.

@DMFTSever I did not touch the Haldane driver here, nor the CDMFT ones
therein, but you may want to do it. I don't really know if this could
have practical implications on what we have been doing (I suspect not)
but for the future I think is a careful move.